feat: replace EvaluationCoefficients with dynamic Map<String,Integer> based on template roles

+++ api/src/main/java/com/system_gestion_soutenance/api/admin/defensesession/dto/CreateDefenseSessionRequest.java
@@ -13,7 +13,7 @@ public record CreateDefenseSessionRequest(
         int defenseDuration,
         int breakDuration,
         String submissionDeadline,
-        Map<String, Double> evaluationCoefficients,
+        Map<String, Integer> evaluationCoefficients,
         String juryRoleTemplateId,
         @NotBlank String startDate,
         @NotBlank String endDate
